Refactor workflow_buidling_modes so it can be imported in dynamic_options.py
